Kinect calibrartion move package
To calibrate the kinekt2 mounted to the head, the pr2 moves a
checkerboard with his left arm to random poses in front of the kinect.

The launch file kinect2_calibration_steps provides the whole calibration
including the start of the callibration on the head.

There is also a demo mode to move the arm to infinite possible calibration
positions without communicating with the head.