tams pr2 simulation initial commit
current state:
The controllers of the pr2 and the shadow hand are managed seperately.
The right arm with the shadow hand falls down before the controllers are loaded
which leads to a not visilbe collision of the first finger and the middle finger.
It is possible to free the collision with the rqt_joint_trajectory_controller.
Without a collision the hand can be moved with moveit.

Changing the max_step_size, the iterations or the sor values of the physic can improve
the simulation.