turn off gravity for the whole right hand
The right arm is falling down when its extended even with loaded controllers.
We determine that the simulation of the motors is to weak. To work around that
problem we are using the same way it is simulated on the left arm.

The "floating" of the left arm is simulated by turning off the gravity for the whole arm.
To get the same behavior for the right arm we need to turn off the gravity too.
It might be necessary to turn it on for the fingers to grasp something.