revert broken MSA configuration
moveit_sim_hw_interface is only useful for moveit_sim_controller,
but breaks if these parameters are not specified

ros_controllers.yaml is not useful in move_group.launch and combines too many things.
It should be outside the "move_group/" namespace for the hardware_interface/controller
to read the parameters, but the "controllers" list belongs in move_group as it specifies
moveit's execution backend.

worst of all, no essential ros_controllers are specified yet, only the joint_state_controller.
So in order to use *any* trajectory execution with ros_control, we still need to specify
these controllers for ros_control and spawn them too.