add adapters for pilz pipeline
technically they will yield invalid trajectories now because
the time parameterization is not included, but the pilz planners
generate their own time parameterization and we mustn't change it.

If we run into problems with this code, we need to add parameterization
ourselves for the Fix* adapters (possibly remove all but FixStartStateBounds again then).