systematically use antiwindup in calibration
This implies the *internally maintained integral term* is clamped
instead of just the output used in the pid controller.
This should have been the default from the start in my opinion, but it wasn't.

This lead to issues with joints hanging in a limit during calibration before
so let's just activate it for all calibration controllers.